+++ components/MenuList/MenuList.js
@@ -1,0 +1,10 @@
+// @flow
+import * as React from 'react';
+import PropTypes from 'prop-types';
+import styles from './Menu.module.scss';
+
+const MenuList = (props: { children: React.Node }) => (
+  <div className={styles.menuList}>{props.children}</div>
+);
+
+export default MenuList;