+++ components/MenuList/MenuSeperator.js
@@ -1,3 +1,4 @@
+// @flow
 import React from 'react';
 import PropTypes from 'prop-types';
 import styles from './Menu.module.scss';