+++ guide/src/components/kebab/index.js
@@ -1,2 +1,0 @@
-import Kebab from './Kebab.js';
-export default Kebab;